aider: Moved the logger to be a property of the `CustomWebSocketConfig` class.
# Aider chat conversation:

USER: the logs still dont appear in the console

ASSISTANT: If the logs are still not appearing, it could be due to the logger being created inside the `registerWebSocketHandlers` method. The logger might be getting garbage collected before it has a chance to print the logs.

To fix this, we can move the logger to be a property of the `CustomWebSocketConfig` class. This way, the logger will exist for the lifetime of the `CustomWebSocketConfig` object and should be able to print the logs.

Here's how you can modify your `WebSocketConfig.kt` file:

src/main/kotlin/com/translate/server/infrastructure/config/WebSocketConfig.kt
```kotlin
<<<<<<< SEARCH
class CustomWebSocketConfig : WebSocketConfigurer {
    override fun registerWebSocketHandlers(registry: WebSocketHandlerRegistry) {
        val handler = customWebSocketHandler()
        val path = "/api/ws"
        registry.addHandler(handler, path)
        val logger = LoggerFactory.getLogger(TranslationController::class.java)
        logger.info("Registered WebSocket handler: ${handler::class.java.name} at path: $path")
    }
=======
class CustomWebSocketConfig : WebSocketConfigurer {
    private val logger = LoggerFactory.getLogger(TranslationController::class.java)

    override fun registerWebSocketHandlers(registry: WebSocketHandlerRegistry) {
        val handler = customWebSocketHandler()
        val path = "/api/ws"
        registry.addHandler(handler, path)
        logger.info("Registered WebSocket handler: ${handler::class.java.name} at path: $path")
    }
>>>>>>> REPLACE
```

This change moves the logger to be a property of the `CustomWebSocketConfig` class.